Use uint64 instead of int32 as IDs for aiContext

+++ Source/abci/Importer/AlembicImporter.cpp
@@ -30,7 +30,7 @@ abciAPI void aiClearContextsWithPath(const char *path)
     aiContextManager::destroyContextsWithPath(path);
 }
 
-abciAPI aiContext* aiContextCreate(int uid)
+abciAPI aiContext* aiContextCreate(uint64_t uid)
 {
     return aiContextManager::getContext(uid);
 }
